requirements.txt
faiss-cpu
requests
streamlit
toml
urllib3
sentence-transformers
pydrive2